fix: unaccessible OTP input
